+++ .github/workflows/ci.yml
@@ -17,5 +17,5 @@ jobs:
           required-ros-distributions: humble
       - uses: ros-tooling/action-ros-ci@v0.3
         with:
-          package-name: rqt_gauges_2
+          package-name: rqt_gauges
           target-ros2-distro: humble
