+++ resource/speedometer.ui
@@ -205,13 +205,13 @@ p, li { white-space: pre-wrap; }
   <customwidget>
    <class>SpeedometerGauge</class>
    <extends>QWidget</extends>
-   <header>rqt_gauges_2.speedometer_gauge.h</header>
+   <header>rqt_gauges.speedometer_gauge.h</header>
    <container>1</container>
   </customwidget>
   <customwidget>
    <class>TopicQLineEdit</class>
    <extends>QLineEdit</extends>
-   <header>rqt_gauges_2.topic_q_line_edit.h</header>
+   <header>rqt_gauges.topic_q_line_edit.h</header>
   </customwidget>
  </customwidgets>
  <resources/>
