+++ setup.cfg
@@ -1,4 +1,4 @@
 [develop]
-script-dir=$base/lib/rqt_gauges_2
+script-dir=$base/lib/rqt_gauges
 [install]
-install-scripts=$base/lib/rqt_gauges_2
+install-scripts=$base/lib/rqt_gauges
