+++ setup.py
@@ -1,6 +1,6 @@
 from setuptools import setup
 
-package_name = 'rqt_gauges_2'
+package_name = 'rqt_gauges'
 
 setup(
     name=package_name,
@@ -23,13 +23,13 @@ setup(
     maintainer='Eloy Briceno',
     maintainer_email='eloy.briceno@ekumenlabs.com',
     description=(
-        'rqt_gauges_2 is a Python GUI plugin providing a visualization tool for several sensors.'
+        'rqt_gauges is a Python GUI plugin providing a visualization tool for several sensors.'
     ),
     license='BSD Clause 3',
     tests_require=['pytest'],
     entry_points={
         'console_scripts': [
-            'rqt_gauges_2 = ' + package_name + '.main:main',
+            'rqt_gauges = ' + package_name + '.main:main',
         ],
     },
 )
