+++ src/rqt_gauges/speedometer.py
