+++ src/rqt_gauges/speedometer_gauge.py
