+++ src/rqt_gauges/steering_wheel_gauge.py
